+++ src/elf/relocator/powerpc64le.rs
@@ -74,7 +74,7 @@ fn resolve_section(
 
     let get_symbol_value = |r: &Relocation| -> Result<u64, LinkerError> {
         match &symbols[r.symbol_index] {
-            ResolvedSymbol::VirtualAddress(v) => Ok(*v as u64),
+            ResolvedSymbol::VirtualAddress(v) => Ok(*v),
             ResolvedSymbol::Absolute(v) => Ok(*v),
             _ => Err(LinkerError::Message(format!(
                 "Symbol at index {} in module {} can not be used for relocation",
@@ -98,12 +98,11 @@ fn resolve_section(
 
         let symbol_value = get_symbol_value(relocation)?;
         let target = symbol_value.wrapping_add(addend as u64);
-
         let place = target_fragment_section.virtual_address + placeholder_offset;
         let signed = match relocation_type {
             RelocationType::R_PPC64_TOC16_HA
             | RelocationType::R_PPC64_TOC16_LO
-            | RelocationType::R_PPC64_TOC16_LO_DS => target.wrapping_sub(toc_value as u64),
+            | RelocationType::R_PPC64_TOC16_LO_DS => target.wrapping_sub(toc_value ),
             _ => target,
         };
 
@@ -125,7 +124,7 @@ fn resolve_section(
             RelocationType::R_PPC64_ADDR64 => PatchItem::from_u64(placeholder_offset, target),
             RelocationType::R_PPC64_REL24 => {
                 let ins = read32(binary, placeholder_offset);
-                let displacement = target.wrapping_sub(place as u64);
+                let displacement = target.wrapping_sub(place );
                 PatchItem::from_u32(
                     placeholder_offset,
                     (ins & !0x03ff_fffc) | ((displacement as u32) & 0x03ff_fffc),
@@ -151,6 +150,10 @@ fn resolve_section(
     Ok(patch_items)
 }
 
-fn read32(binary: &[u8], offset: usize) -> u32 {
-    u32::from_le_bytes(binary[offset..offset + 4].try_into().unwrap())
+fn read32(binary: &[u8], offset: u64) -> u32 {
+    u32::from_le_bytes(
+        binary[offset as usize..offset as usize + 4]
+            .try_into()
+            .unwrap(),
+    )
 }
